Finished first pass at the key unwrap

+++ cpp/libs/secauthv5/src/secauthv5/outstation/KeyChangeState.cpp
@@ -33,7 +33,7 @@ using namespace opendnp3;
 namespace secauthv5
 {
 
-	KeyChangeState::KeyChangeState(uint16_t userNum, uint8_t challengeSize_, openpal::Logger logger_, openpal::ICryptoProvider& provider) :
+	KeyChangeState::KeyChangeState(uint16_t userNum, uint16_t challengeSize_, openpal::Logger logger_, openpal::ICryptoProvider& provider) :
 		USER_NUM(userNum),
 		challengeSize(AuthConstants::GetBoundedChallengeSize(challengeSize_)),
 		logger(logger_),
